Split network module out from eCFR parser
