+++ solution/parser/ecfr-parser/go.mod
@@ -4,7 +4,9 @@ go 1.16
 
 require (
 	github.com/aws/aws-lambda-go v1.27.1
-	github.com/go-test/deep v1.0.8 // indirect
-	github.com/sirupsen/logrus v1.8.1
-	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616 // indirect
+	github.com/cmsgov/cmcs-eregulations/network v0.0.0-00010101000000-000000000000
+	github.com/go-test/deep v1.0.8
+	github.com/sirupsen/logrus v1.9.0
 )
+
+replace github.com/cmsgov/cmcs-eregulations/network => ../network
