+++ solution/parser/fr-parser/eregs/eregs.go
@@ -9,7 +9,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/cmsgov/cmcs-eregulations/ecfr-parser/network"
+	"github.com/cmsgov/cmcs-eregulations/network"
 
 	log "github.com/sirupsen/logrus"
 )
