+++ solution/parser/fr-parser/fedreg/fedreg.go
@@ -10,7 +10,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/cmsgov/cmcs-eregulations/ecfr-parser/network"
+	"github.com/cmsgov/cmcs-eregulations/network"
 
 	log "github.com/sirupsen/logrus"
 )
