+++ solution/parser/network/go.mod
@@ -1,0 +1,8 @@
+module github.com/cmsgov/cmcs-eregulations/network
+
+go 1.16
+
+require (
+	github.com/go-test/deep v1.0.8
+	github.com/sirupsen/logrus v1.9.0
+)
