+++ solution/parser/network/network.go
