+++ solution/parser/network/network_test.go
